moving functions to schema-functions file
