+++ packages/destination-actions/src/destinations/hubspot/customEvent/utils.ts
@@ -1,6 +1,5 @@
 import { PayloadValidationError, IntegrationError, RetryableError, StatsContext } from '@segment/actions-core'
 import { SubscriptionMetadata } from '@segment/actions-core/destination-kit'
-import type { Payload } from './generated-types'
 import {
   CreateEventDefinitionReq,
   CreatePropDefinitionReq,
@@ -19,36 +18,6 @@ export const cache = new LRUCache<string, CachableSchema>({
   max: 2000,
   ttl: 1000 * 60 * 60
 })
-
-function stringFormat(str: string): StringFormat {
-  // Check for date or datetime, otherwise default to string
-  const isoDateTimeRegex =
-    /^(\d{4})-(0[1-9]|1[0-2])-(0[1-9]|[12]\d|3[01])(?:([ T])(\d{2}):?(\d{2})(?::?(\d{2})(?:[,\.](\d{1,}))?)?(?:(Z)|([+\-])(\d{2})(?::?(\d{2}))?)?)?$/ //eslint-disable-line no-useless-escape
-  const dateOnlyRegex = /^\d{4}-\d{2}-\d{2}$/ //eslint-disable-line no-useless-escape
-
-  if (isoDateTimeRegex.test(str)) {
-    return dateOnlyRegex.test(str) ? 'date' : 'datetime'
-  } else {
-    return 'string'
-  }
-}
-
-export function eventSchema(payload: Payload): Schema {
-  const { event_name, properties } = payload
-  const props: { [key: string]: SegmentProperty } = {}
-
-  if (properties) {
-    Object.entries(properties).forEach(([property, value]) => {
-      if (value !== null) {
-        props[property] = {
-          type: typeof value as SegmentPropertyType,
-          stringFormat: typeof value === 'string' ? stringFormat(value) : undefined
-        }
-      }
-    })
-  }
-  return { name: event_name, primaryObject: payload.record_details.object_type, properties: props }
-}
 
 export function getSchemaFromCache(
   name: string,
